[Refactor] API uri 변경
